# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Property AI Studio (PropVista CRM) API
  version: 1.0.0
  description: 'REST API for PropVista CRM. Base path /api/v1. JWT Bearer auth. Gemini-only
    AI. Roles: customer, agent, admin, super_admin. Guest = no auth. Errors: {error:{code,message,details[]}}.
    Pagination meta. Prices as strings. Notification channels: email, in_app.'
servers:
- url: http://localhost:4001/api/v1
  description: Local development
- url: https://staging-api.propvista.example/api/v1
  description: Staging
- url: https://api.propvista.example/api/v1
  description: Production
...
paths:
...
  /agents:
    get:
      tags:
      - Agents
      summary: List agents
      description: 'Auth: JWT Bearer. Authorization: customer, agent, admin, super_admin.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: getAgents
      security:
      - bearerAuth: []
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
    post:
      tags:
      - Agents
      summary: Create agent
      description: 'Auth: JWT Bearer. Authorization: admin, super_admin.'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'CONFLICT_DUPLICATE_EMAIL'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'VALIDATION_ERROR'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: postAgents
      security:
      - bearerAuth:
        - admin
        - super_admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
  /agents/{id}:
    get:
      tags:
      - Agents
      summary: Get agent
      description: 'Auth: JWT Bearer. Authorization: customer, agent, admin, super_admin.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'RESOURCE_NOT_FOUND'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: getAgentById
      security: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Resource identifier
    patch:
      tags:
      - Agents
      summary: Update agent
      description: 'Auth: JWT Bearer. Authorization: admin, super_admin, agent.'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'RESOURCE_NOT_FOUND'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'VALIDATION_ERROR'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: patchAgentById
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Resource identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
    delete:
      tags:
      - Agents
      summary: Delete agent
      description: 'Auth: JWT Bearer. Authorization: admin, super_admin.'
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_INVALID_CREDENTIALS'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'AUTH_FORBIDDEN'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: 'RESOURCE_NOT_FOUND'
                  message: Human-readable error message
                  details:
                  - field: example
                    issue: validation failed
      operationId: deleteAgentById
      security:
      - bearerAuth:
        - admin
        - super_admin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Resource identifier
  /agents/{id}/image:
    post:
      tags:
      - Agents
      summary: Upload agent profile image
      description: 'Auth: JWT Bearer. Authorization: admin, super_admin. Local disk storage in development.'
      operationId: postAgentImage
      security:
      - bearerAuth:
        - admin
        - super_admin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Invalid type or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - details
          properties:
            code:
              type: string
              description: Stable application error code (SRS Appendix P)
              enum:
              - AUTH_INVALID_CREDENTIALS
              - AUTH_TOKEN_EXPIRED
              - AUTH_FORBIDDEN
              - VALIDATION_ERROR
              - RESOURCE_NOT_FOUND
              - CONFLICT_DUPLICATE_EMAIL
              - LEAD_IDEMPOTENCY_REPLAY
              - AI_UNAVAILABLE
              - AI_TIMEOUT
              - BULK_PARSE_ERROR
              - BULK_VALIDATION_FAILED
              - PROPERTY_PUBLISH_BLOCKED
              - RATE_LIMITED
              - INTERNAL_ERROR
            message:
              type: string
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
    Meta:
      type: object
      required:
      - page
      - pageSize
      - total
      - totalPages
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    Agent:
      type: object
      required:
      - id
      - name
      - email
      - isActive
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        profileImageUrl:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentCreate:
      type: object
      required:
      - name
      - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        userId:
          type: string
          format: uuid
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        userId:
          type: string
          format: uuid
          nullable: true
        isActive:
          type: boolean
...
security:
- bearerAuth: []